+++ Server/Migrations/SQLite/20240214112349_InitialCreateSQLite.cs
@@ -27,7 +27,7 @@ namespace LCPCollection.Server.Migrations.SQLite
                     ReleaseDate = table.Column<DateTime>(type: "TEXT", nullable: true),
                     TrailerUrl = table.Column<string>(type: "TEXT", nullable: true),
                     Genre = table.Column<string>(type: "TEXT", nullable: true),
-                    Rating = table.Column<double>(type: "REAL", nullable: true)
+                    Rating = table.Column<float>(type: "REAL", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -50,7 +50,7 @@ namespace LCPCollection.Server.Migrations.SQLite
                     ReleaseDate = table.Column<DateTime>(type: "TEXT", nullable: true),
                     TrailerUrl = table.Column<string>(type: "TEXT", nullable: true),
                     Genre = table.Column<string>(type: "TEXT", nullable: true),
-                    Rating = table.Column<double>(type: "REAL", nullable: true)
+                    Rating = table.Column<float>(type: "REAL", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -91,7 +91,7 @@ namespace LCPCollection.Server.Migrations.SQLite
                     ReleaseDate = table.Column<DateTime>(type: "TEXT", nullable: true),
                     TrailerUrl = table.Column<string>(type: "TEXT", nullable: true),
                     Genre = table.Column<string>(type: "TEXT", nullable: true),
-                    Rating = table.Column<double>(type: "REAL", nullable: true)
+                    Rating = table.Column<float>(type: "REAL", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -114,7 +114,7 @@ namespace LCPCollection.Server.Migrations.SQLite
                     ReleaseDate = table.Column<DateTime>(type: "TEXT", nullable: true),
                     TrailerUrl = table.Column<string>(type: "TEXT", nullable: true),
                     Genre = table.Column<string>(type: "TEXT", nullable: true),
-                    Rating = table.Column<double>(type: "REAL", nullable: true)
+                    Rating = table.Column<float>(type: "REAL", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -135,7 +135,7 @@ namespace LCPCollection.Server.Migrations.SQLite
                     Version = table.Column<int>(type: "INTEGER", nullable: true),
                     Requirements = table.Column<string>(type: "TEXT", nullable: true),
                     Platforms = table.Column<string>(type: "TEXT", nullable: true),
-                    Rating = table.Column<double>(type: "REAL", nullable: true),
+                    Rating = table.Column<float>(type: "REAL", nullable: true),
                     DateCreated = table.Column<DateTime>(type: "TEXT", nullable: true),
                     AuthorName = table.Column<string>(type: "TEXT", nullable: true),
                     UrlValue = table.Column<string>(type: "TEXT", nullable: true)
@@ -161,7 +161,7 @@ namespace LCPCollection.Server.Migrations.SQLite
                     ReleaseDate = table.Column<DateTime>(type: "TEXT", nullable: true),
                     TrailerUrl = table.Column<string>(type: "TEXT", nullable: true),
                     Genre = table.Column<string>(type: "TEXT", nullable: true),
-                    Rating = table.Column<double>(type: "REAL", nullable: true)
+                    Rating = table.Column<float>(type: "REAL", nullable: true)
                 },
                 constraints: table =>
                 {
@@ -181,7 +181,7 @@ namespace LCPCollection.Server.Migrations.SQLite
                     ImageUrl = table.Column<string>(type: "TEXT", nullable: true),
                     Requirements = table.Column<string>(type: "TEXT", nullable: true),
                     BrowsersName = table.Column<string>(type: "TEXT", nullable: true),
-                    Rating = table.Column<double>(type: "REAL", nullable: true),
+                    Rating = table.Column<float>(type: "REAL", nullable: true),
                     DateCreated = table.Column<DateTime>(type: "TEXT", nullable: true),
                     AuthorName = table.Column<string>(type: "TEXT", nullable: true),
                     UrlValue = table.Column<string>(type: "TEXT", nullable: true)
@@ -194,7 +194,7 @@ namespace LCPCollection.Server.Migrations.SQLite
             migrationBuilder.InsertData(
                 table: "Animes",
                 columns: new[] { "Id", "Companies", "CoverUrl", "Description", "Genre", "ImageUrl", "Platforms", "Publishers", "Rating", "ReleaseDate", "Title", "TrailerUrl" },
-                values: new object[] { 1, "Toei Animation", "covers/db.jpg", "Dragon Ball", "Animation", "images/db.jpg", "TV", "Toei Animation", 9.0, new DateTime(1986, 2, 26, 0, 0, 0, 0, DateTimeKind.Utc), "Dragon Ball", "https://www.youtube.com/watch?v=gqIEgmqljM8" });
+                values: new object[] { 1, "Toei Animation", "covers/db.jpg", "Dragon Ball", "Animation", "images/db.jpg", "TV", "Toei Animation", 9f, new DateTime(1986, 2, 26, 0, 0, 0, 0, DateTimeKind.Utc), "Dragon Ball", "https://www.youtube.com/watch?v=gqIEgmqljM8" });
         }
 
         /// <inheritdoc />
